Code cleanup and modularization.  Testing script created.
