+++ scripts/test.sh
@@ -32,8 +32,21 @@ echo ""
 #python validate_article.py -D --color=tty -f ../CuratedContent/TestDrivenDevptInScientificSwASurvey.md
 
 
+#echo "================================================================================"
+#echo ""
+# python validate_article.py -D --color=tty -f ../CuratedContent/BestPracticesForHPCSwDevelopersWebinarSeries.md
+
+
 echo "================================================================================"
 echo ""
-python validate_article.py -D --color=tty -f ../CuratedContent/BestPracticesForHPCSwDevelopersWebinarSeries.md
+python validate_article.py -f test_file_001.md -s metadata_spec.txt -D --color=tty
+
+echo "================================================================================"
+echo ""
+python validate_article.py -f test_file_002.md -s metadata_spec.txt -D --color=tty
+
+echo "================================================================================"
+echo ""
+python validate_article.py -f test_file_003.md -s metadata_spec.txt -D --color=tty
